feat: Updated category picker to V2
The existing categories picker now contains a 'New category' option, which if selected shows the 'New Category' text input field on the form. This will reduce clutter on the form and aid users in selecting a new OR existing category. Future updates to be worked on to allow users to select multiple categories and to edit categories on existing habits.