feat: Added ability to move between data set views on charts
The addition of a left and right chevron either side of reports allows users to navigate back and forth to view all available data. If the date chunk they are looking at contains the earliest data point for a habit, they cannot navigate further back, vice versa for the latest date point and navigating forwards.

Still to be implemented - adjusting data labels and data for weekly and monthly views. Currently just shows data in 7 day chunks for daily habits, 4 day chunks for monthly, and 12 day chunks for year. To be developed.